requirements.txt
requests>=2.26.0
urllib3>=1.26.9
sqlalchemy>=1.4.40
beautifulsoup4>=4.11.1
PyYAML>=6.0
apscheduler
redis
pytest